+++ reeln_meta_plugin/__init__.py
@@ -2,7 +2,7 @@
 
 from __future__ import annotations
 
-__version__ = "0.5.0"
+__version__ = "0.7.0"
 
 from reeln_meta_plugin.plugin import MetaPlugin